+++ Package.swift
@@ -1,5 +1,5 @@
 // swift-tools-version: 5.9
-// FlinkuSDK 0.3.0
+// FlinkuSDK 0.3.1
 import PackageDescription
 
 let package = Package(